linting and move figuremaker function
