+++ lineage/states/StateDistPhase.py
@@ -1,14 +1,16 @@
-from math import gamma
+""" State distribution class for separated G1 and G2 phase durations as observation. """
 import numpy as np
 import scipy.stats as sp
-from numba import njit
-import scipy.special as sc
-from scipy.optimize import brentq
-from .StateDistribution import gamma_estimator, gamma_pdf
-from .stateCommon import bern_pdf, bernoulli_estimator
+from .StateDistribution import (gamma_estimator,
+                                gamma_pdf,
+                               )
+from .stateCommon import (bern_pdf,
+                          bernoulli_estimator,
+                         )
 
 
 class StateDistribution2:
+    """ For G1 and G2 separated as observations. """
     def __init__(self, bern_p, gamma_a1, gamma_scale1, gamma_a2, gamma_scale2):  # user has to identify what parameters to use for each state
         """ Initialization function should take in just in the parameters for the observations that comprise the multivariate random variable emission they expect their data to have. """
         self.bern_p = bern_p
@@ -82,14 +84,14 @@ class StateDistribution2:
         # if this function runs again.
         return state_estimate_obj
 
-    def tHMM_E_init(self):
-        """
-        Initialize a default state distribution.
-        """
-        return StateDistribution2(0.9, 7, 3 + (1 * (np.random.uniform())), 14, 6 + (1 * (np.random.uniform())))
-
     def __repr__(self):
         """
         Method to print out a state distribution object.
         """
         return "State object w/ parameters: {}, {}, {}.".format(self.bern_p, self.gamma_a1, self.gamma_scale1, self.gamma_a2, self.gamma_scale2)
+
+def tHMM_E_init(self):
+    """
+    Initialize a default state distribution.
+    """
+    return StateDistribution2(0.9, 7, 3 + (1 * (np.random.uniform())), 14, 6 + (1 * (np.random.uniform())))
